requirements.txt
einops
scikit-learn==1.0.0
packaging
numpy==1.19.5
fastparquet==0.7.1
hydra-core
matplotlib
h5py
nlb_tools
pandas==1.3.4
tensorboard==2.11.0
wandb
pandas
xformers==0.0.27.post2
flash-attn==2.6.3
kaggle==1.6.17
zenodo_get
unzip